refactor(nef): extract scanner engine into analyze module
Move the syntax-tree walker, analyze_file_at, FileInfo, and all the
AST/resolver helpers into scan/analyze.rs. mod.rs keeps only the public
API (CatalogRef, scan_package, scan_package_with_roots) plus the tests.
FileInfo, analyze_file_at, identity_origins, and line_col are pub(super)
for the graph module and tests; everything else stays private. The
module is named for analyze_file_at, its entry point; the Walker is an
internal detail.
